Refactor Header and ManageWorkLayout components; integrate API data fetching for modules, enhance navigation logic, and conditionally render components based on module availability
